eHealth-backend # Changes: - refactor code, - create service-discovery
