+++ device-service/src/main/java/com/mstrzezon/deviceservice/controller/DeviceController.java
@@ -1,4 +1,4 @@
-package com.mstrzezon.deviceservice.repository;
+package com.mstrzezon.deviceservice.controller;
 
 import com.mstrzezon.deviceservice.dto.DeviceRequest;
 import com.mstrzezon.deviceservice.dto.DeviceResponse;
