+++ device-service/src/main/resources/application.properties
@@ -1,2 +1,5 @@
 spring.data.mongodb.uri=mongodb+srv://michal:UQruBVa80GwpxItm@cluster0.2wlc9t1.mongodb.net/devices?retryWrites=true&w=majority
 spring.data.mongodb.database=devices
+
+eureka.client.service-url.defaultZone=http://localhost:8761/eureka
+spring.application.name=device-service
