+++ discovery-server/src/main/resources/application.properties
@@ -1,0 +1,4 @@
+eureka.instance.hostname=localhost
+eureka.client.register-with-eureka=false
+eureka.client.fetch-registry=false
+server.port=8761
