+++ sensorsdata-service/src/main/resources/application.properties
@@ -1,7 +1,9 @@
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
-spring.datasource.url=jdbc:mysql://localhost:3306/sensorsdata-service
-spring.datasource.username=admin
-spring.datasource.password=admin
+spring.datasource.url=jdbc:mysql://mysql.agh.edu.pl:3306/mstrzezo
+spring.datasource.username=mstrzezo
+spring.datasource.password=Z5zYTvajwNsHtFgq
 spring.jpa.hibernate.ddl-auto=update
 
 server.port=8081
+eureka.client.service-url.defaultZone=http://localhost:8761/eureka
+spring.application.name=sensorsdata-service
